// client/src/types/Token.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TokenBar } from "./TokenBar";

export type Token = { id: string, layer_id: string, name: string, asset_id: string | null, owner_id: string | null, x: number, y: number, size: number, rotation: number, bars: Array<TokenBar>, status_markers: Array<string>, has_vision: boolean, vision_range: number, darkvision_range: number, light_bright: number, light_dim: number, created_at: string, updated_at: string, };